Refactor tracking - extract from repo
