+++ src/WebApp.Core/app/components/login-page.ts
@@ -3,11 +3,13 @@ import * as ko from "knockout";
 import { EventTypes } from "../models/EventTypes";
 import { repositoryFactory } from "../repositories/repositoryFactory";
 import { config } from "../config";
+import { trackingFactory } from "../repositories/trackingFactory";
 
 export class LoginPageViewModel {
   currentComponent = ko.observable().syncWith("currentComponent", true, true);
   currentCustomer = ko.observable().syncWith("currentCustomer", true, true);
   repository = repositoryFactory.get();
+  tracking = trackingFactory.get();
 
   constructor() {}
 
@@ -17,11 +19,7 @@ export class LoginPageViewModel {
     response.then(data => {
       this.currentCustomer(data);
 
-      this.repository.trackEvent(
-        data,
-        EventTypes.userLogin,
-        "Visitor logged in"
-      );
+      this.tracking.trackEvent(data, EventTypes.userLogin, "Visitor logged in");
 
       this.currentComponent("start-page");
     });
